requirements.txt
flask
gspread
google-auth
python-dotenv
requests
pandas
numpy